requirements.txt
requests
aiohttp
pyjwt
python-dotenv
pytest-aiohttp
pylint
coverage
cchardet
aiodns
asyncio